refactor(tab): using the latest tab component with typescript
